% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{check_mslevels}
\alias{check_mslevels}
\title{Extract data from Msnobject}
\usage{
check_mslevels(data)
}
\arguments{
\item{data}{MSnObject or a data.table with (file, rt, mz, i)}
}
\value{
An integer with the available MSlevels
}
\description{
Extract data from Msnobject
}
