% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{extract_XICs}
\alias{extract_XICs}
\title{Extract XICs}
\usage{
extract_XICs(roi_ls, rttol = 0, debugL = FALSE)
}
\arguments{
\item{roi_ls}{roi list obtained with \code{ionshed()}}

\item{rttol}{Tolerance in rt dimension (grid points)}

\item{debugL}{Boolean to print debug messages}
}
\value{
A \code{data.table} object with:
\itemize{
\item \verb{$xic_roi} \code{INTEGER}: roi extracted (XIC id)
\item \verb{$rt} \code{NUMERIC}: retention time
\item \verb{$mz} \code{NUMERIC}: m/Z
\item \verb{$i} \code{NUMERIC}: intensity
\item \verb{$roi} \code{INTEGER}: original roi
\item \verb{$assignment} \code{LOGICAL}: is point inside current roi (\code{xic_roi})
}
}
\description{
Extract XICs
}
