% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{fun_create_map_single}
\alias{fun_create_map_single}
\title{Extract signals coordinates and map on a sparse grid}
\usage{
fun_create_map_single(
  data,
  file_ind = 1,
  mz_res = 5e-05,
  rt_res = 0.1,
  mslevels = 1,
  isowin_mz = NULL,
  rt_range = NULL
)
}
\arguments{
\item{data}{OnDiskMSnExp, MSnObject, MsExperiment or a data.table
with (file, rt, mz, i)}

\item{file_ind}{File index to subset}

\item{mz_res}{m/Z resolution in Dalton to create the grid}

\item{rt_res}{rt resolution in second to create the grid}

\item{mslevels}{choose msLevels to extract (default to mslevel=1)}

\item{isowin_mz}{\code{numeric} value to filter a specific isolation window}

\item{rt_range}{\code{numeric} vector \code{c(rtmin, rtmax)} to define the rt
range to extract. Default \code{NULL}}
}
\value{
A \code{list} with:
\itemize{
\item \code{resolution}: a \code{list} with \code{mz} and \code{rt} resolutions
\item \code{data}: a \code{data.table} with:
\itemize{
\item \code{idmz}: m/Z index
\item \code{idrt}: rt index
\item \code{i}: intensity
\item \code{rt}: retention time in seconds
\item \code{mz}: m/Z
}
}
}
\description{
Extract signals coordinates and map on a sparse grid
}
\examples{
fun_create_map_single(data = MSnbase::filterFile(msnexp_obj, 1), file_ind = 1)
}
