% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_builder.R
\name{get_xcmsobj}
\alias{get_xcmsobj}
\title{Create XCMS object with ionshed peaks}
\usage{
get_xcmsobj(peak_info, data, mslevel = 1L)
}
\arguments{
\item{peak_info}{peak info obtained from \code{ionshed()}}

\item{data}{OnDiskMSnExp, MSnObject, MsExperiment or a data.table
with (file, rt, mz, i)}

\item{mslevel}{MS level to assign peaks in the xcms object}
}
\value{
A \verb{xcms object} with manual peak integration.
}
\description{
Create XCMS object with ionshed peaks
}
