% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed-package.R
\docType{package}
\name{ionshed-package}
\alias{ionshed-package}
\title{ionshed: Ionshed segmentation of LCMS data}
\description{
Sparse matrix segmentation inspired by watershed and implemented in C++..
}
\author{
\strong{Maintainer}: Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr} (\href{https://orcid.org/0000-0002-3387-7250}{ORCID})

Authors:
\itemize{
  \item Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr} (\href{https://orcid.org/0000-0002-3387-7250}{ORCID})
}

Other contributors:
\itemize{
  \item Etienne Thevenot \email{etienne.thevenot@cea.fr} (\href{https://orcid.org/0000-0003-1019-4577}{ORCID}) [contributor]
}

}
\keyword{internal}
