% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{ionshed}
\alias{ionshed}
\title{Get roi and peaks from MSnobject}
\usage{
ionshed(
  data,
  file_ind = 1,
  mz_res = 0.001,
  rt_res = 0.01,
  rttol = 5,
  ppm = 5,
  subset_dt = NULL,
  mslevels = 1,
  rt_range = NULL,
  by_isowin = TRUE,
  debugL = FALSE
)
}
\arguments{
\item{data}{MSnObject or a data.table with (file, rt, mz, i)}

\item{file_ind}{File index to subset}

\item{mz_res}{m/Z resolution in Dalton to create the grid}

\item{rt_res}{rt resolution in second to create the grid}

\item{rttol}{Tolerance in rt dimension (grid points)}

\item{ppm}{Tolerance in ppm (variable grid)}

\item{subset_dt}{An optional \code{data.table} to restrict search space with rtmin, rtmax, mzmin, mzmax}

\item{mslevels}{choose msLevels to extract (default to mslevel=1)}

\item{by_isowin}{\code{Logical} to run extraction for each isolation windows separately (DIA)}

\item{debugL}{Boolean to print debug messages}
}
\value{
A \code{list} with:
\itemize{
\item \code{roi_data}: A \code{data.table} with:
\itemize{
\item \code{rt_sc}: retention time index
\item \code{mz_sc}: m/Z index
\item \code{i}: intensity
\item \code{roi}: roi index
\item \code{roiType}: roi type
}
\item \code{peak_info}: A \code{data.table} with:
\itemize{
\item \code{roi}: roi index
\item \code{sc_length}: segment lenght
\item \code{rtid_start}: retention time index starting location
\item \code{rtid_end}: retention time index ending location
\item \code{mzid_start}: m/Z index starting location
\item \code{mzid_end}: m/Z index ending location
\item \code{rtmean}: retention time mean in seconds
\item \code{rtmin}: retention time minimum value in seconds
\item \code{rtmax}: retention time maximum value in seconds
\item \code{mzmin}: m/Z minimum value
\item \code{mzmax}: m/Z maximum value
\item \code{rtmed}: retention time median in seconds
\item \code{rtmeanw}: retention intensity weighted mean in seconds
\item \code{mzmed}: m/Z median value
\item \code{mzmeanw}: m/Z intensity weighted mean
\item \code{imax}: intensity maximum value
\item \code{imin}: intensity minimum value
\item \code{idiff}: intensity difference between max and min
\item \code{imedian}: intensity median value
\item \code{imean}: intensity mean
\item \code{into}: sum of intensities inside the segment
\item \code{mzdiff}: m/Z maximum difference
\item \code{rtdiff}: retention time maximum difference in seconds
}
}
}
\description{
Get roi and peaks from MSnobject
}
\examples{
res <- ionshed(MSnobject = MSnExpObj, file_ind = 1, rttol = 10, ppm = 10)
str(res)
}
