% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSnExpObj}
\alias{MSnExpObj}
\title{This is a sample dataset to run example}
\format{
An object of class \code{MSnExp} with 3 rows and 1 columns.
}
\usage{
MSnExpObj
}
\description{
This dataset correspond to a truncated MSnExp object
acquired on LC-MS (Orbitrap) system in ESI+.
}
\references{
\url{https://www.mdpi.com/2218-1989/10/4/158}
}
\keyword{datasets}
