% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpMat_DescendingROI}
\alias{SpMat_DescendingROI}
\title{Function to attribute ROI by taking the highest intensity points}
\usage{
SpMat_DescendingROI(
  locations,
  values,
  rttol = 10L,
  ppm = 10,
  debug = FALSE,
  eachiter = FALSE
)
}
\arguments{
\item{locations}{A matrix (2xN) with row (1xN) and column (1xN)}

\item{values}{A vector of N values}

\item{rttol}{Tolerance in rt dimension (grid points)}

\item{ppm}{Tolerance in ppm (variable grid)}

\item{debug}{Boolean to print debug messages}

\item{eachiter}{Boolean to return iteration steps details}
}
\description{
This function returns a matrix with m/Z groups
based on a tolerance. This version is the final one.
}
