% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zigzag_index}
\alias{zigzag_index}
\title{Function to compute the ZigZag index}
\usage{
zigzag_index(intPts)
}
\arguments{
\item{intPts}{Numeric vector of intensity pre-ordered by time}
}
\description{
This function returns the ZigZag index as a numeric
value of a single trace.
}
