% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{check_mslevels}
\alias{check_mslevels}
\title{Check available ms levels}
\usage{
check_mslevels(data)
}
\arguments{
\item{data}{OnDiskMSnExp, MSnObject, MsExperiment or a data.table
with (file, rt, mz, i)}
}
\value{
An integer with the available MSlevels
}
\description{
Check available ms levels
}
