% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{fun_create_map_single}
\alias{fun_create_map_single}
\title{Extract data from Msnobject}
\usage{
fun_create_map_single(
  data,
  file_ind = 1,
  mz_res = 5e-05,
  rt_res = 0.1,
  mslevels = 1,
  isowin_mz = NULL,
  rt_range = NULL
)
}
\arguments{
\item{data}{MSnObject or a data.table with (file, rt, mz, i)}

\item{file_ind}{File index to subset}

\item{mz_res}{m/Z resolution in Dalton to create the grid}

\item{rt_res}{rt resolution in second to create the grid}

\item{mslevels}{choose msLevels to extract (default to mslevel=1)}
}
\value{
A \code{list} with:
\itemize{
\item \code{resolution}: a \code{list} with \code{mz} and \code{rt} resolutions
\item \code{data}: a \code{data.table} with:
\itemize{
\item \code{idmz}: m/Z index
\item \code{idrt}: rt index
\item \code{i}: intensity
\item \code{rt}: retention time in seconds
\item \code{mz}: m/Z
}
}
}
\description{
Extract data from Msnobject
}
\examples{
data_in <- MSnbase::filterFile(MSnExpObj, 1)
fun_create_map_single(data_in, file_ind = 1)
}
