% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{get_rawdata}
\alias{get_rawdata}
\title{Get raw data from MSnexp object}
\usage{
get_rawdata(data_in)
}
\arguments{
\item{data_in}{MSnExp MsExperiment or XcmsExperiment object}
}
\value{
A \code{data.table} object with:
\itemize{
\item \code{file}: file number
\item \code{scan}: scan index
\item \code{rt}: retention time in seconds
\item \code{mz}: m/Z
\item \code{i}: Intensity
}
}
\description{
Get raw data from MSnexp object
}
\examples{
data_in <- MSnbase::filterFile(MSnExpObj, 1)
get_rawdata(data_in)
}
