% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionshed.R
\name{get_rawdata}
\alias{get_rawdata}
\title{Get raw data from MSnexp object}
\usage{
get_rawdata(data)
}
\arguments{
\item{data}{OnDiskMSnExp, MSnObject, MsExperiment or a data.table
with (file, rt, mz, i)}
}
\value{
A \code{data.table} object with:
\itemize{
\item \code{file}: file number
\item \code{scan}: scan index
\item \code{rt}: retention time in seconds
\item \code{mz}: m/Z
\item \code{i}: Intensity
}
}
\description{
Get raw data from MSnexp object
}
\examples{
get_rawdata(data = MSnbase::filterFile(msnexp_obj, 1))
}
