% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_builder.R
\name{get_xcmsobj}
\alias{get_xcmsobj}
\title{Create XCMS object with ionshed peaks}
\usage{
get_xcmsobj(peak_info, msnobj, msLevel = 1L)
}
\arguments{
\item{peak_info}{peak info obtained from \code{ionshed()}}
}
\value{
A \verb{xcms object} with manual peak integration.
}
\description{
Create XCMS object with ionshed peaks
}
