% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpMat_getEICs}
\alias{SpMat_getEICs}
\title{Function to extract EIcs using Sparse Matrix}
\usage{
SpMat_getEICs(locations, values, eics_dt, rttol = 10L, debugL = FALSE)
}
\arguments{
\item{locations}{A matrix (2xN) with row (1xN) and column (1xN)}

\item{values}{A matrix (ValuesxN) with values in the same order as locations}

\item{eics_dt}{coordinate as a matrix with id, mzmin, mzmax, rtmin, rtmax}

\item{rttol}{(optional) to extend the rt range}

\item{debugL}{Boolean to print debug message}
}
\description{
This function returns a matrice with EICs corresponding to asked range(s).
}
